feat(sso): expose Connection.type and consolidate ConnectionType
Regenerate SSO with the connection_type wire field surfaced under the
domain name `type` (via the new fieldHints), preserving the long-standing
public SDK surface; wire serialization still uses connection_type. Drop the
redundant hand-written connection-type.enum and point ListConnectionsOptions
at the generated connection-type.interface so there is a single
ConnectionType. Update the events spec's connection fixture to match.
