+++ src/events/events.spec.ts
@@ -63,7 +63,7 @@ describe('Event', () => {
       id: 'conn_01234ABCD',
       organizationId: 'org_1234',
       name: 'Connection',
-      connectionType: ConnectionType.OktaSAML,
+      type: ConnectionType.OktaSAML,
       state: 'active',
       domains: [],
       createdAt: new Date('2020-05-06 04:21:48.649164'),
