--- src/sso/interfaces/connection-type.enum.ts
@@ -1,40 +1,0 @@
-export enum ConnectionType {
-  ADFSSAML = 'ADFSSAML',
-  AdpOidc = 'AdpOidc',
-  AppleOAuth = 'AppleOAuth',
-  Auth0SAML = 'Auth0SAML',
-  AzureSAML = 'AzureSAML',
-  CasSAML = 'CasSAML',
-  CleverOIDC = 'CleverOIDC',
-  ClassLinkSAML = 'ClassLinkSAML',
-  CloudflareSAML = 'CloudflareSAML',
-  CyberArkSAML = 'CyberArkSAML',
-  DuoSAML = 'DuoSAML',
-  EntraIdOIDC = 'EntraIdOIDC',
-  GenericOIDC = 'GenericOIDC',
-  GenericSAML = 'GenericSAML',
-  GitHubOAuth = 'GitHubOAuth',
-  GoogleOAuth = 'GoogleOAuth',
-  GoogleSAML = 'GoogleSAML',
-  JumpCloudSAML = 'JumpCloudSAML',
-  KeycloakSAML = 'KeycloakSAML',
-  LastPassSAML = 'LastPassSAML',
-  LoginGovOidc = 'LoginGovOidc',
-  MagicLink = 'MagicLink',
-  MicrosoftOAuth = 'MicrosoftOAuth',
-  MiniOrangeSAML = 'MiniOrangeSAML',
-  NetIqSAML = 'NetIqSAML',
-  OktaOIDC = 'OktaOIDC',
-  OktaSAML = 'OktaSAML',
-  OneLoginSAML = 'OneLoginSAML',
-  OracleSAML = 'OracleSAML',
-  PingFederateSAML = 'PingFederateSAML',
-  PingOneSAML = 'PingOneSAML',
-  RipplingSAML = 'RipplingSAML',
-  SalesforceOAuth = 'SalesforceOAuth',
-  SalesforceSAML = 'SalesforceSAML',
-  ShibbolethGenericSAML = 'ShibbolethGenericSAML',
-  ShibbolethSAML = 'ShibbolethSAML',
-  SimpleSamlPhpSAML = 'SimpleSamlPhpSAML',
-  VMwareSAML = 'VMwareSAML',
-}
