+++ src/sso/interfaces/connection-type.interface.ts
@@ -4,23 +4,30 @@ export enum ConnectionType {
   ADFSSAML = 'ADFSSAML',
   AdpOidc = 'AdpOidc',
   AppleOAuth = 'AppleOAuth',
+  Auth0Migration = 'Auth0Migration',
   Auth0SAML = 'Auth0SAML',
   AzureSAML = 'AzureSAML',
+  BitbucketOAuth = 'BitbucketOAuth',
   CasSAML = 'CasSAML',
   ClassLinkSAML = 'ClassLinkSAML',
   CleverOIDC = 'CleverOIDC',
   CloudflareSAML = 'CloudflareSAML',
   CyberArkSAML = 'CyberArkSAML',
+  DiscordOAuth = 'DiscordOAuth',
   DuoSAML = 'DuoSAML',
   EntraIdOIDC = 'EntraIdOIDC',
   GenericOIDC = 'GenericOIDC',
   GenericSAML = 'GenericSAML',
   GitHubOAuth = 'GitHubOAuth',
+  GitLabOAuth = 'GitLabOAuth',
   GoogleOAuth = 'GoogleOAuth',
+  GoogleOidc = 'GoogleOIDC',
   GoogleSAML = 'GoogleSAML',
+  IntuitOAuth = 'IntuitOAuth',
   JumpCloudSAML = 'JumpCloudSAML',
   KeycloakSAML = 'KeycloakSAML',
   LastPassSAML = 'LastPassSAML',
+  LinkedInOAuth = 'LinkedInOAuth',
   LoginGovOidc = 'LoginGovOidc',
   MagicLink = 'MagicLink',
   MicrosoftOAuth = 'MicrosoftOAuth',
@@ -30,6 +37,7 @@ export enum ConnectionType {
   OktaSAML = 'OktaSAML',
   OneLoginSAML = 'OneLoginSAML',
   OracleSAML = 'OracleSAML',
+  Pending = 'Pending',
   PingFederateSAML = 'PingFederateSAML',
   PingOneSAML = 'PingOneSAML',
   RipplingSAML = 'RipplingSAML',
@@ -38,17 +46,9 @@ export enum ConnectionType {
   ShibbolethGenericSAML = 'ShibbolethGenericSAML',
   ShibbolethSAML = 'ShibbolethSAML',
   SimpleSamlPhpSAML = 'SimpleSamlPhpSAML',
-  VMwareSAML = 'VMwareSAML',
-  Pending = 'Pending',
-  Auth0Migration = 'Auth0Migration',
-  BitbucketOAuth = 'BitbucketOAuth',
-  DiscordOAuth = 'DiscordOAuth',
-  GitLabOAuth = 'GitLabOAuth',
-  GoogleOidc = 'GoogleOIDC',
-  IntuitOAuth = 'IntuitOAuth',
-  LinkedInOAuth = 'LinkedInOAuth',
   SlackOAuth = 'SlackOAuth',
   TestIdp = 'TestIdp',
+  VMwareSAML = 'VMwareSAML',
   VercelMarketplaceOAuth = 'VercelMarketplaceOAuth',
   VercelOAuth = 'VercelOAuth',
   XeroOAuth = 'XeroOAuth',
